sjauth: Use curl's default print-to-stdout, rather than writing to a buffer
I noticed that printf was frequently reading past the end of the
response because the write function doesn't place a null terminator when
it writes to the response buffer. Instead of figuring out the length of
the response, we could simply use curl's default CURLOPT_WRITEFUNCTION
and the default CURLOPT_WRITEDATA, i.e. writing to stdout.